Cleanup script enhancements
These changes are made to reflect the AWS image builder security best
practices documented here:
https://docs.aws.amazon.com/imagebuilder/latest/userguide/security-best-practices.html

There were a few files that were not deleted in the script above, so at
the end of the new cleanup script we will continue deleting those with
'rm' and will engage with AWS teams about updating the image builder
script to include those files.

Our script also does not allow configuration of each individual deletion
stage. This is not needed for the OSS ami recipes since users could just
customize this script to their liking if they are creating a custom AMI.