Add include_deprecated=true to source_ami_filter
During an AMI build, Packer uses the name of the Amazon Linux source AMI defined in release-{ami-family}.auto.pkvars.hcl
It uses the most recent AMI that has the given name. In AL1 recipe, we added included_deprecated=true since the AL1 base AMIs are deprecated.

This commit adds the same flag to all other recipes, so that whenever the base AMI is deprecated, Packer is able to still use it.