Add ecs-log-collector to AMI builds (excludes AL1)
ECS Log Collector is a script for gathering logs to troubleshooting
Amazon ECS customer support cases. In some cases it is difficult to
download the script into customer EC2 instance(s).

This change aims to include the script as part of the AMI build process
to ensure that we remove one obstacle from customer's gathering logs for
troubleshooting.

The version of the script is currently pinned to a specific commit hash
to allow the team control over what version is included with an AMI
release. The script and version file can be observed in the
/opt/amazon/ecs folder:
- ecs-logs-collector.sh - ECS Logs Collector shell script
- ECS_LOG_COLLECTOR_VERSION - Contains the commit hash used to pull the
  ECS Logs Collector shell script