disable containerd's CRI plugin on ECS AL2023 AMIs
The CRI plugin is an implementation of Kubernetes container runtime interface (CRI). The cri plugin inside containerd handles all CRI service requests from the Kubelet and uses containerd internals to manage containers and container images.

The CRI plugin is not relevant in ECS world, hence this commit disables it in the ECS Optimized AMIs.